+++ src/components/home_page/props.ts
@@ -1,5 +1,5 @@
-import { TaskListItem } from '../../selectors/tasks';
-import { ExploreAllProps }  from '../explore/explore_all';
+import { TaskListItem } from '../../selectors/tasks/task_list_item';
+import { ExploreAllProps } from '../explore/explore_all';
 import { Id as TaskId } from '../../stores/tasks';
 
 export interface HomePageProps extends ExploreAllProps {
