+++ src/components/tasks/related_tasks.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text } from 'native-base';
 import { applicationStyles } from '../../application/styles';
-import { TaskListItem } from '../../selectors/tasks';
+import { TaskListItem } from '../../selectors/tasks/task_list_item';
 import { TaskListComponent, noTasksAddedYetTextComponent } from './task_list';
 import { Trans } from '@lingui/react';
 import { TaskListItemActions } from './task_list_item';
