+++ src/selectors/articles/article.ts
@@ -1,5 +1,5 @@
 import * as model from '../../stores/articles';
-import { TaskListItem } from '../tasks';
+import { TaskListItem } from '../tasks/task_list_item';
 import { ExploreSection } from '../explore/types';
 import { ArticleListItem } from './article_list_item';
