+++ src/selectors/tasks/task_list_item.ts
@@ -1,0 +1,7 @@
+export interface TaskListItem {
+    readonly id: string;
+    readonly title: string;
+    readonly description: string;
+    readonly isRecommended: boolean;
+    readonly completed: boolean;
+}
