+++ src/selectors/tasks/to_selector_task.ts
@@ -3,7 +3,7 @@ import { getLocalizedText } from '../locale/get_localized_text';
 import { Locale } from '../../locale/types';
 import { ArticleListItem } from '../articles/article_list_item';
 import { ExploreSection } from '../explore/types';
-import { TaskListItem } from '.';
+import { TaskListItem } from './task_list_item';
 import { Task } from './task';
 
 export const toSelectorTask =
